Working on storing the configuration in higher memory.
Changes to project to allow memory access above the 16 bits into the 17 bits area for data structures.

Still not happy with the final memory allocation.

Adding extra configuration and number of executions to make sure the data is being stored correctly.

Considering adding a checksum to the config to compensate for data corruption on the field.

Starting work on adding the calibration inside the main function and run it on press at boot.